Added stub testcase ('not implemented') for missing tests
Depending on ICS feature selection some test cases do not apply. Such test
cases that do not apply for the implemented entry point have not been
implemented. With this fix stub testcases are implemented for EPs that
implement a different feature set. Only a warning is printed and the respective
stub testcase fails.